Added meetups-store.js and moved meetups array along with the toggleFavorite and addMeetup functions inside
